# scripts/bridge.py
"""Generate the strict transport bridge, preserving forward-compatible JSON fields."""
import json
import os
from pathlib import Path
os.chdir(Path(__file__).resolve().parent.parent)
spec=json.load(open('api/openapi.yaml'))
s='''// Code generated by scripts/bridge.py; DO NOT EDIT.
package http

import (
 "context"
 "encoding/json"
 "net/http"
 "github.com/team-screaner/screener-service/api"
 "github.com/team-screaner/screener-service/internal/domain"
)
type jsonResponse struct { data domain.Result }
func (r jsonResponse) write(w http.ResponseWriter) error {
 w.Header().Set("Content-Type","application/json")
 w.WriteHeader(http.StatusOK)
 return json.NewEncoder(w).Encode(r.data)
}
'''
for path,ops in spec['paths'].items():
 for method,op in ops.items():
  name=op['operationId'];g=name[0].upper()+name[1:]
  if name=='exportXLSX':continue
  resource='request.Id' if '{id}' in path else '""'
  s+=f'func (r jsonResponse) Visit{g}Response(w http.ResponseWriter)error{{return r.write(w)}}\n'
  s+=f'func (s *server) {g}(ctx context.Context, request api.{g}RequestObject)(api.{g}ResponseObject,error){{ result,err:=s.execute(ctx,"{name}",{resource});if err!=nil{{return nil,err}};return jsonResponse{{data:result}},nil }}\n'
Path('internal/adapter/http/bridge.gen.go').write_text(s)
# Generated artifacts are committed formatted; generation must be byte-for-byte reproducible.
import subprocess
subprocess.run(['gofmt','-w','internal/adapter/http/bridge.gen.go'],check=True)